double --> float transition

+++ inc/point.h
@@ -7,19 +7,19 @@ namespace Raytracer
     class Point
     {
         public:
-            Point(double x, double y, double z, double w);
-            Point(double x, double y, double z);
+            Point(float x, float y, float z, float w);
+            Point(float x, float y, float z);
             virtual ~Point();
-            double x() const;
-            double y() const;
-            double z() const;
-            double w() const;
+            float x() const;
+            float y() const;
+            float z() const;
+            float w() const;
             Point operator+(const Vector v) const;
             Point &operator+=(const Vector v);
             Vector operator-(const Point p) const;
             friend std::ostream& operator<<(std::ostream& out, Point p);
         protected:
-            double mX, mY, mZ, mW;
+            float mX, mY, mZ, mW;
     };
 }
 #endif
